+++ include/analysis.hpp
@@ -253,6 +253,27 @@ namespace MathLib {
         }
 
         /**
+         * @brief Calculates arc tangent of input value "x", using 7 degree-polynomial approximation.
+         * "x" must be in the range [-1, 1], otherwise the calculation error increases.
+         * @details The coefficients a1,a3,...,a7 were calculated as they minimize the cost function f(a1,a3,...,a7) := \int_0^1 (a1*x + a3*x^3 + ... + a7*x^7 - atan(x))^2 \mathrm{d}x.
+         * The maximal absolute error is less than 1.8*10^(-4) when 0<=x<=1.
+         *
+         * @tparam T the number type of the input value
+         * @param[in] x input value
+         * @return arc tangent of "x"
+         */
+        template <typename T>
+        T atan_polyApprox_deg7(T x) {
+            static_assert(std::is_floating_point<T>::value, "argument type must be floating point number.");
+            constexpr T a1 = 0.999298;
+            constexpr T a3 = -0.322084;
+            constexpr T a5 = 0.148508;
+            constexpr T a7 = -0.0404899;
+            const T y = x*x;
+            return x*(a1 + y*(a3 + y*(a5 + a7*y)));
+        }
+
+        /**
          * @brief Calculates arc tangent of input value "x", using 9 degree-polynomial approximation.
          * "x" must be in the range [-1, 1], otherwise the calculation error increases.
          * @details The 5 coefficients a1,a3,...,a9 were calculated as they minimize the cost function f(a1,a3,...,a9) := \int_0^1 (a1*x + a3*x^3 + ... + a9*x^9 - atan(x))^2 \mathrm{d}x.
@@ -282,24 +303,51 @@ namespace MathLib {
         }
 
         /**
-         * @brief Calculates arc tangent of input value "x", using 7 degree-polynomial approximation.
-         * "x" must be in the range [-1, 1], otherwise the calculation error increases.
-         * @details The coefficients a1,a3,...,a7 were calculated as they minimize the cost function f(a1,a3,...,a7) := \int_0^1 (a1*x + a3*x^3 + ... + a7*x^7 - atan(x))^2 \mathrm{d}x.
-         * The maximal absolute error is less than 1.8*10^(-4) when 0<=x<=1.
+         * @brief Calculates "atan2(y,x)" using polynomial approximation (internally calls `atan_polyApprox_deg7` function).
          *
          * @tparam T the number type of the input value
-         * @param[in] x input value
-         * @return arc tangent of "x"
+         * @param y "y"
+         * @param x "x"
+         * @return "atan2(y,x)"
          */
         template <typename T>
-        T atan_polyApprox_deg7(T x) {
+        T atan2_polyApprox_deg7(T y, T x) {
             static_assert(std::is_floating_point<T>::value, "argument type must be floating point number.");
-            constexpr T a1 = 0.999298;
-            constexpr T a3 = -0.322084;
-            constexpr T a5 = 0.148508;
-            constexpr T a7 = -0.0404899;
-            const T y = x*x;
-            return x*(a1 + y*(a3 + y*(a5 + a7*y)));
+            constexpr T POSITIVE_ZERO = +0.0;
+            constexpr T NEGATIVE_ZERO = -0.0;
+            constexpr T pi = 3.14159265358979323846;
+            constexpr T half_pi = pi/2;
+
+            if (x == NEGATIVE_ZERO || x == POSITIVE_ZERO) {
+                if (y >= POSITIVE_ZERO) {
+                    return half_pi;
+                }
+                if (y <= NEGATIVE_ZERO) {
+                    return -half_pi;
+                }
+            }
+            if (x > POSITIVE_ZERO) {
+                if (-x <= y && y <= x) {
+                    return atan_polyApprox_deg7(y/x);
+                }
+                if (y > x) {
+                    return half_pi - atan_polyApprox_deg7(x/y);
+                } else { // y < -x
+                    return -half_pi - atan_polyApprox_deg7(x/y);
+                }
+            } else { // x < NEGATIVE_ZERO
+                if (POSITIVE_ZERO <= y && y <= -x) {
+                    return pi + atan_polyApprox_deg7(y/x);
+                }
+                if (-x < y) {
+                    return half_pi - atan_polyApprox_deg7(x/y);
+                }
+                if (x <= y && y <= NEGATIVE_ZERO) {
+                    return atan_polyApprox_deg7(y/x) - pi;
+                } else { // y < x
+                    return -half_pi - atan_polyApprox_deg7(x/y);
+                }
+            }
         }
 
         /**