+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Analysis::atan_polyApprox_deg7` function.
 - Renamed `Analysis::atan_polyApprox` to `Analysis::atan_polyApprox_deg9`.
 - Renamed `Analysis::atan2_polyApprox` to `Analysis::atan2_polyApprox_deg9`.
+- Added `Analysis::atan2_polyApprox_deg7` function.
 
 ## [0.4.0] - 2021/8/27